Parts 6-8
Added ad package with basics of ad management classes for the restaurant.
AdvertisementStorage is ment to be only one for the whole restaurant, so it will be a Singleton.

Initialized AdvertisementStorage videos field with new List and added few "ads" in constructor of that class.
Added storage field to AdManager class

Added getters to Advertisement method