Part 10
processVideos() method:
It had to present videos that have highest amount of money per impression, secondary sort is by video length (the more, the better) and tertiary by video count in the set (the less the better).

It used recursion to process all videos.

Bugfix: ConsoleHelper class - ensured that exception is not thrown when entering empty string.

Advertisement class - ensured that if we get duration or impressionsRemaining passed to constructor, equal to 0, we won't get divideby0 exceptions in several methods.